versione 2.0

+++ src/main/scala/com/umana/corso/catalogue/domain/model/Movie.scala
@@ -1,3 +1,2 @@
 package com.umana.corso.catalogue.domain.model
-
-case class Movie(id: String, producer: String, title: String,duration:Int,price:Double,category:String)
+case class Movie(id: String, producer: String, title: String,duration:Int,price:BigDecimal,category:String)
